moved all sagas into folder and created a search GET and a remove favorite DELETE
