+++ src/index.js
@@ -5,73 +5,20 @@ import { legacy_createStore as createStore, combineReducers, applyMiddleware } f
 
 import { Provider } from 'react-redux';
 import logger from 'redux-logger';
-import { takeEvery, put } from "redux-saga/effects";
 import createSagaMiddleware from "redux-saga";
+import _rootSaga from './redux/sagas/_root.saga';
 
 
 const sagaMiddleware = createSagaMiddleware();
 
 
-function* watcherSaga() {
-  yield takeEvery("FETCH_CATEGORY", getCategorySaga);
-  yield takeEvery("ADD_CATEGORY", addCategorySaga);
-  yield takeEvery("FETCH_FAVORITE", getFavoriteSaga);
-  yield takeEvery("ADD_FAVORITE", addFavoriteSaga);
-}
+// function* watcherSaga() {
+//   yield takeEvery("FETCH_CATEGORY", getCategorySaga);
+//   yield takeEvery("ADD_CATEGORY", addCategorySaga);
+//   yield takeEvery("FETCH_FAVORITE", getFavoriteSaga);
+//   yield takeEvery("ADD_FAVORITE", addFavoriteSaga);
+// }
 
-function* getCategorySaga(){
-    try {
-        const response = yield axios({
-          method: "GET",
-          url: "/api/category",
-        });
-        yield put({ type: "SET_CATEGORY", payload: response.data });
-      } catch (error) {
-        console.log("error with GET request", error);
-      }
-}
-
-function* addCategorySaga(action) {
-    try {
-      yield axios({
-        method: "POST",
-        url: "/api/category",
-        data: { name: action.payload.name },
-      });
-  
-      yield put({ type: "FETCH_CATEGORY" });
-    } catch (error) {
-      console.log(error);
-      alert("Unable to add category");
-    }
-  }
-
-  function* getFavoriteSaga(){
-    try {
-        const response = yield axios({
-          method: "GET",
-          url: "/api/favorite",
-        });
-        yield put({ type: "SET_FAVORITE", payload: response.data });
-      } catch (error) {
-        console.log("error with GET request", error);
-      }
-}
-
-function* addFavoriteSaga(action) {
-    try {
-      yield axios({
-        method: "POST",
-        url: "/api/favorite",
-        data: { name: action.payload.name },
-      });
-  
-      yield put({ type: "FETCH_FAVORITES" });
-    } catch (error) {
-      console.log(error);
-      alert("Unable to add category");
-    }
-  }
 
 const storeInstance = createStore(
     combineReducers({
@@ -80,7 +27,7 @@ const storeInstance = createStore(
     applyMiddleware(sagaMiddleware, logger),
 );
 
-sagaMiddleware.run(watcherSaga);
+sagaMiddleware.run(_rootSaga);
 
 ReactDOM.render(
 <Provider store={storeInstance}>
